Verificando se os arquivos temporarios de certificado ainda existem ao destruir a classe

+++ src/Soap/SoapBase.php
@@ -177,11 +177,7 @@ abstract class SoapBase implements SoapInterface
      */
     public function __destruct()
     {
-        try {
-            $this->removeTemporarilyFiles();
-        }
-        catch (\Exception $e) {
-        }
+        $this->removeTemporarilyFiles();
     }
 
     /**
@@ -543,9 +539,18 @@ abstract class SoapBase implements SoapInterface
             return;
         }
         //remove os certificados
-        $this->filesystem->delete($this->certfile);
-        $this->filesystem->delete($this->prifile);
-        $this->filesystem->delete($this->pubfile);
+        if ($this->filesystem->has($this->certfile)) {
+            $this->filesystem->delete($this->certfile);
+        }
+
+        if ($this->filesystem->has($this->prifile)) {
+            $this->filesystem->delete($this->prifile);
+        }
+
+        if ($this->filesystem->has($this->pubfile)) {
+            $this->filesystem->delete($this->pubfile);
+        }
+
         //remove todos os arquivos antigos
         $contents = $this->filesystem->listContents($this->certsdir, true);
         $dt = new \DateTime();
@@ -555,7 +560,7 @@ abstract class SoapBase implements SoapInterface
         foreach ($contents as $item) {
             if ($item['type'] == 'file') {
                 $timestamp = $this->filesystem->getTimestamp($item['path']);
-                if ($timestamp < $tsLimit) {
+                if ($timestamp < $tsLimit && $this->filesystem->has($item['path'])) {
                     $this->filesystem->delete($item['path']);
                 }
             }
